tests: separated helper functions into standalone file
